rename js files & add new index.js files to comp folder
